Change prefix symbol and move etherwarp into visuals package directly
